feat: add basic rendering
