--- .husky/pre-push
@@ -1,4 +1,0 @@
-#!/usr/bin/env sh
-
-. "$(dirname -- "$0")/_/husky.sh"
- npx vitest run
