+++ src/app/(main)/list/layout.tsx
@@ -1,0 +1,7 @@
+export default function MainPageLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <div>{children}</div>;
+}
