+++ src/app/(main)/list/page.tsx
@@ -1,0 +1,5 @@
+import './styles/main-page.css';
+
+export default function MainPage() {
+  return <div className="card-container">List </div>;
+}
