+++ src/app/(main)/list/styles/main-page.css
@@ -1,12 +1,3 @@
-.body {
-  box-sizing: border-box;
-  padding: 20px;
-
-  background-color: #213547;
-  background-image: url('/img/space.jpeg');
-  background-size: cover;
-}
-
 .main-content {
   margin: 100px 50px;
 }
