--- src/app/list/page.tsx
@@ -1,13 +1,0 @@
-import './main-page.css';
-import './search-bar.css';
-import './header.css';
-
-export default function MainPage() {
-  return (
-    <div className="main-wrapper" data-testid="main-screen">
-      <div className="wrapper-outlet">
-        <div className="card-container">List </div>
-      </div>
-    </div>
-  );
-}
